+++ javascript/extraNetworks.js
@@ -19,6 +19,7 @@ function setupExtraNetworksForTab(tabname) {
     searchTerm = search.value.toLowerCase();
     gradioApp().querySelectorAll(`#${tabname}_extra_tabs div.card`).forEach((elem) => {
       text = `${elem.querySelector('.name').textContent.toLowerCase()} ${elem.querySelector('.search_term').textContent.toLowerCase()}`;
+      text = text.replace('models--', 'Diffusers')
       elem.style.display = text.indexOf(searchTerm) == -1 ? 'none' : '';
     });
   });