+++ modules/sd_hijack.py
@@ -174,14 +174,12 @@ class StableDiffusionModelHijack:
             sd_hijack_unet.hijack_ddpm_edit()
 
         if opts.cuda_compile and opts.cuda_compile_mode == 'ipex':
-            import logging
             if shared.cmd_opts.use_ipex:
                 shared.log.info("Model compile enabled: IPEX Optimize Graph Mode")
             else:
                 shared.log.warning("Model compile skipped: IPEX Method is for Intel GPU's with OneAPI")
         elif opts.cuda_compile and opts.cuda_compile_mode != 'none' and shared.backend == shared.Backend.ORIGINAL:
             try:
-                import logging
                 import torch._dynamo as dynamo # pylint: disable=unused-import
                 # torch._dynamo.config.log_level = logging.WARNING if opts.cuda_compile_verbose else logging.CRITICAL # pylint: disable=protected-access
                 torch._dynamo.config.verbose = opts.cuda_compile_verbose # pylint: disable=protected-access