+++ modules/shared.py
@@ -6,12 +6,10 @@ import datetime
 import urllib.request
 from urllib.parse import urlparse
 from enum import Enum
-import tempfile
 import gradio as gr
 import tqdm
 import requests
-import diffusers
-from modules import errors, ui_components, shared_items, cmd_args, modelloader
+from modules import errors, ui_components, shared_items, cmd_args
 from modules.paths_internal import models_path, script_path, data_path, sd_configs_path, sd_default_config, sd_model_file, default_sd_model_file, extensions_dir, extensions_builtin_dir # pylint: disable=W0611
 import modules.interrogate
 import modules.memmon
@@ -235,48 +233,6 @@ def list_checkpoint_tiles():
 default_checkpoint = list_checkpoint_tiles()[0] if len(list_checkpoint_tiles()) > 0 else "model.ckpt"
 
 
-def load_diffusers_lora(lora_repo: str):
-    pipe = sys.modules[__name__].sd_model
-    if lora_repo == "":
-        pipe._remove_text_encoder_monkey_patch() # pylint: disable=W0212
-        proc_cls_name = next(iter(pipe.unet.attn_processors.values())).__class__.__name__
-        non_lora_proc_cls = getattr(diffusers.models.attention_processor, proc_cls_name[len("LORA"):])
-        pipe.unet.set_attn_processor(non_lora_proc_cls())
-        return ""
-    elif len(lora_repo.split('/')) == 2:
-        lora_dir = os.path.dirname(opts.data["diffusers_dir"])
-        cache_dir = os.path.join(lora_dir, "Diffusers_LoRA")
-        pipe.load_lora_weights(lora_repo, cache_dir=cache_dir)
-        print(f"Loaded {lora_repo}")
-        return f"{lora_repo} is loaded. Pass empty text field to remove LoRA or pass new LoRA id."
-    else:
-        print(f"{lora_repo} is not a valid LoRA identifier.")
-        return ""
-
-
-def load_diffusers_text_inv(text_inv_repo: str):
-    pipe = sys.modules[__name__].sd_model
-    if text_inv_repo == "":
-        pipe.tokenizer = pipe.tokenizer.__class__.from_pretrained(pipe.tokenizer.name_or_path)
-        pipe.text_encoder.resize_token_embeddings(len(pipe.tokenizer))
-        return ""
-    elif is_url(text_inv_repo):
-        with tempfile.TemporaryDirectory() as temp_dir:
-            os.system(f"wget -P {temp_dir} {text_inv_repo}")
-            temp_file_path = os.path.join(temp_dir, text_inv_repo.split('/')[-1])
-            pipe.load_textual_inversion(temp_file_path)
-            text_inv_repo = '/'.join(text_inv_repo.split('/')[-2:])
-        print(f"Loaded Civit.ai Textual Inv: {text_inv_repo}")
-    elif len(text_inv_repo.split('/')) == 2:
-        text_inv_dir = os.path.dirname(opts.data["diffusers_dir"])
-        cache_dir = os.path.join(text_inv_dir, "Diffusers_Text_Inv")
-        pipe.load_textual_inversion(text_inv_repo, cache_dir=cache_dir)
-        print(f"Loaded {text_inv_repo}")
-    text_inv_tokens = pipe.tokenizer.added_tokens_encoder.keys()
-    text_inv_tokens = [t for t in text_inv_tokens if not (len(t.split("_")) > 1 and t.split("_")[-1].isdigit())]
-    return f"{', '.join(text_inv_tokens)} loaded. Pass empty text field to remove all or add new textual inversion id."
-
-
 def refresh_checkpoints():
     import modules.sd_models # pylint: disable=W0621
     return modules.sd_models.list_models()
@@ -349,7 +305,7 @@ options_templates.update(options_section(('sd', "Stable Diffusion"), {
     "prompt_mean_norm": OptionInfo(True, "Prompt attention mean normalization"),
     "comma_padding_backtrack": OptionInfo(20, "Prompt padding for long prompts", gr.Slider, {"minimum": 0, "maximum": 74, "step": 1 }),
     "sd_disable_ckpt": OptionInfo(False, "Disallow usage of checkpoints in ckpt format"),
-    "sd_backend": OptionInfo("Original", "Stable Diffusion backend (experimental)", gr.Radio, lambda: {"choices": ["Original", "Diffusers"] }),
+    "sd_backend": OptionInfo("original", "Stable Diffusion backend (experimental)", gr.Radio, lambda: {"choices": ["original", "diffusers"] }),
 }))
 
 options_templates.update(options_section(('optimizations', "Optimizations"), {
@@ -483,7 +439,6 @@ options_templates.update(options_section(('ui', "User Interface"), {
     "ui_tab_reorder": OptionInfo("From Text, From Image, Process Image", "UI tabs order"),
     "ui_scripts_reorder": OptionInfo("Enable Dynamic Thresholding, ControlNet", "UI scripts order"),
     "ui_reorder": OptionInfo(", ".join(ui_reorder_categories), "txt2img/img2img UI item order"),
-    "ui_extra_networks_tab_reorder": OptionInfo("Checkpoints, Lora, LyCORIS, Textual Inversion, Hypernetworks", "Extra networks tab order"),
 }))
 
 options_templates.update(options_section(('live-preview', "Live Previews"), {
@@ -573,11 +528,13 @@ options_templates.update(options_section(('upscaling', "Upscaling"), {
 }))
 
 options_templates.update(options_section(('extra_networks', "Extra Networks"), {
+    "ui_extra_networks_tab_reorder": OptionInfo("Checkpoints, Lora, LyCORIS, Textual Inversion, Hypernetworks", "Extra networks tab order"),
     "extra_networks_card_cover": OptionInfo("inline", "UI position", gr.Radio, lambda: {"choices": ["cover", "inline", "sidebar"]}),
     "extra_networks_sidebar_width": OptionInfo(35, "UI sidebar width (%)", gr.Slider, {"minimum": 10, "maximum": 80, "step": 1}),
     "extra_networks_card_size": OptionInfo(200, "UI card size (px)", gr.Slider, {"minimum": 20, "maximum": 2000, "step": 1}),
     "extra_networks_card_square": OptionInfo(False, "UI disable variable aspect ratio"),
     "extra_networks_card_fit": OptionInfo("cover", "UI image contain method", gr.Radio, lambda: {"choices": ["contain", "cover", "fill"]}),
+    "extra_network_skip_indexing": OptionInfo(False, "Do not automatically build extra network pages", gr.Checkbox),
     "lyco_patch_lora": OptionInfo(False, "Use LyCoris handler for all Lora types", gr.Checkbox),
     "lora_disable": OptionInfo(False, "Disable built-in Lora handler", gr.Checkbox, { "visible": True }, onchange=lora_disable),
     "lora_functional": OptionInfo(False, "Use Kohya method for handling multiple Loras", gr.Checkbox),
@@ -734,13 +691,10 @@ opts = Options()
 config_filename = cmd_opts.config
 opts.load(config_filename)
 cmd_opts = cmd_args.compatibility_args(opts, cmd_opts)
-if cmd_opts.backend == 'diffusers':
-    log.info('Overriding backend to Diffusers')
-    opts.data['sd_backend'] = 'Diffusers'
-if cmd_opts.backend == 'original':
-    log.info('Overriding backend to Diffusers')
-    opts.data['sd_backend'] = 'Original'
-backend = Backend.DIFFUSERS if opts.sd_backend == 'Diffusers' else Backend.ORIGINAL
+if cmd_opts.backend:
+    opts.data['sd_backend'] = cmd_opts.backend.lower()
+backend = Backend.DIFFUSERS if opts.sd_backend == 'diffusers' else Backend.ORIGINAL
+log.info(f'Pipeline: {cmd_opts.backend.lower()}')
 
 prompt_styles = modules.styles.StyleDatabase(opts.styles_dir)
 cmd_opts.disable_extension_access = (cmd_opts.share or cmd_opts.listen or (cmd_opts.server_name or False)) and not cmd_opts.insecure